+++ src/PhraseSeq16.cpp
@@ -1249,13 +1249,18 @@ struct PhraseSeq16 : Module {
 				int newSeq = seqIndexEdit;// good value when editingSequence, overwrite if not editingSequence
 				if (ppqnCount == 0) {
 					float slideFromCV = 0.0f;
+					int oldStepIndexRun = stepIndexRun;
 					if (editingSequence) {
 						slideFromCV = cv[seqIndexEdit][stepIndexRun];
-						moveIndexRunMode(&stepIndexRun, sequences[seqIndexEdit].getLength(), sequences[seqIndexEdit].getRunMode(), &stepIndexRunHistory);
+						bool seqLoopOver = moveIndexRunMode(&stepIndexRun, sequences[seqIndexEdit].getLength(), sequences[seqIndexEdit].getRunMode(), &stepIndexRunHistory);
+						if (seqLoopOver && stopAtEndOfSong) {
+							running = false;
+							stepIndexRun = oldStepIndexRun;
+						}
 					}
 					else {
 						slideFromCV = cv[phrase[phraseIndexRun]][stepIndexRun];
-						int oldStepIndexRun = stepIndexRun;
+						
 						if (moveIndexRunMode(&stepIndexRun, sequences[phrase[phraseIndexRun]].getLength(), sequences[phrase[phraseIndexRun]].getRunMode(), &stepIndexRunHistory)) {
 							int oldPhraseIndexRun = phraseIndexRun;
 							bool songLoopOver = moveIndexRunMode(&phraseIndexRun, phrases, runModeSong, &phraseIndexRunHistory);
@@ -1925,7 +1930,7 @@ struct PhraseSeq16Widget : ModuleWidget {
 		holdItem->module = module;
 		menu->addChild(holdItem);
 
-		StopAtEndOfSongItem *loopItem = createMenuItem<StopAtEndOfSongItem>("Stop at end of song", CHECKMARK(module->stopAtEndOfSong));
+		StopAtEndOfSongItem *loopItem = createMenuItem<StopAtEndOfSongItem>("Single shot", CHECKMARK(module->stopAtEndOfSong));
 		loopItem->module = module;
 		menu->addChild(loopItem);